requirements.txt
segment-geospatial 
segment-anything-fast 
matplotlib
Pillow
gradio
